requirements.txt
python-dotenv
pytest
fastapi
uvicorn
passlib
bcrypt
sqlalchemy
psycopg2
alembic
